Use taint copy to prevent variable index overwriting on the same instance.
Happens when mutating multiple variables ... #0,1(etc). Works fine with
one mutatinon. The first valueTaint instance is saved as a variable taint
with the right variable index.
When mutating second variable we cannot reuse the previous taint object
and store as variable taint for 2nd index, we would change the index of
both.