Add new detector for MODIFICATION_AFTER_VALIDATION
A string must not be modified after validation because it may allow an
attacker to bypass validation using a tricky string which becomes
malicious after the modification. For example, a program may filter out
the <script> tags from HTML input to avoid cross-site scripting (XSS) and
other vulnerabilities. If non-character code points are deleted from the
input following validation, an attacker may pass the string "<scr"+
"\uFDEF"+"ipt>" so that the validation check fails to detect the <script>
tag, but the subsequent removal of the non-character code pont creates a
<script> tag in the input.

This patch adds a detector for this bug.