New detector for SEI-CERT rule SEC02-J: Do not base security checks on untrusted sources
A new detector the checks whether the same non-final  method of a non-final
class that comes from a user was called before the privileged action and inside
it. Since neither the class, nor the method is final the user provided class
may override this method to behave in an unexpected, bypassing security checks.