Correctly parse method signatures with generic types
Currently, the implementation assumes the signature will be of the format
`(Ljava/lang/Object,Ljava/lang/Object)`, using a comma `,` separator.
This is not the case for compiled classes. Compiled signatures use a
semicolon `;` as the separator as described in [JVMS 4.2](https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-4.html#jvms-ParameterDescriptor).

This commit delegates the `SpotBugs'` signature parser to ensure the parser
expects the format it is to be provided & reduces the maintenance overhead here.