Update invokedynamic workaround to check the size of parameters like long and double.
The long and double primitive type parameters occupy two stack registers so we check the locations of stack parameters that can modify taint status. Any reference types as well as primitive char will propagate taint to the resulting String.

Do not add a mutable stack index to the TaintMethodConfig because we don't have a mutable object on the stack in the case of the invokedynamic call.