Test cases for invokedynamic workaround for String concatenation.
In order to test invokedynamic string concatenation handling, we need to build some sample code with a later JDK so use Java 11 for the compiler target in a new build submodule. We cannot currently switch the existing findsecbugs-samples-java to Java 11 without breaking tests, and a new module is added just for the Java 11 classes we want to test.

Added setter for FindSecBugsGlobalConfig.workaroundVisitInvokeDynamic that the unit test uses.