Misinterpreting invokedynamic causes "missing class" warning
When analyzing Java bytecode for "invoke" instructions, we assumed that
there always was a valid class name specified when calling
InvokeInstruction#getClassName.

Unfortunately, this is not true for "invokedynamic" calls, where BCEL
repurposes this method to get the method name.

This leads to confusing warnings of the type "The following classes
needed for analysis were missing: makeConcatWithConstants", etc.

Add code that checks for the INVOKEDYNAMIC subclass, skipping these
instructions whenever necessary.

Fixes #332