Remove array signature from classname
When looking up a class name, we sometimes mistake an array type for a
simple classname, causing a subsequent class lookup to fail.

That failure is reported to Spotbugs, but apparently ignored internally.

Instead of ignoring this, we should try to extract the simple class
type, and check this one instead.

Fixes #751